Fix hang when calling functions in a module with an on_load attribute from a native module
The fix consists in adding a new opcode for hipe stub calls that
is used to replace the call to error_handler:undefined_function/3. This
stub directly calls the function: there is no need to go through
error_handler:undefined_function/3 once we know the module is loaded,
and, more importantly, when error_handler module itself was natively
compiled (with native code loaded), the emulator did loop infinitely.

The fix also comes with a non-regression test that will prevent this
bug to reappear again.